requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
pandas==2.2.0
reportlab==4.0.9
Pillow==10.2.0
python-dotenv==1.0.1
django-environ==0.11.2
psycopg2-binary==2.9.9  # For PostgreSQL support if needed
gunicorn==21.2.0  # For production deployment
firebase-admin==6.4.0  # For FCM notifications 